requirements.txt
aiohttp==3.8.4
redis==5.0.1
uvloop==0.17.0; platform_system != "Windows"
uvicorn==0.21.1
pydantic==1.10.10
pytest==7.4.0
locust==2.18.0
prometheus_client==0.16.0
python-dotenv==1.0.
aiohttp-cors
